Added various keywords to BNGLScanner.java, including "seed", "functions", "DeleteMolecules", etc. Also removed all actions from the keyword list, such as "generate_network", "simulate_ode", "writeSBML", etc. Did this primarily because there are a lot of actions and new actions are being added often. Keeping the keyword list up to date will thus be difficult. Second, actions are not really part of the model, so it makes sense for them to be colored differently from model elements.
One small issue is that "seed" and "actions", which have been added to the keyword list, are also names of arguments for the simulate() and generate_hybrid_model() actions, respectively. As such, they get colored differently from other action arguments. This is a minor issue, but it would be nice to eventually differentiate model keywords from action arguements having the same name.