Short term fix for console confusion.
Essentially, writing to the console requires a string id for the console, but any unrecognized id was just spawning a new console.  This flexibility is great for not caring whether or not you need to initialize a console, but was causing multiple consoles to be created for the same model since variations of the id were being used (relative path to to file, absolute path to file, relative path with no leading slash, relative path with leading slash).  Everything now consistently uses relative paths with no leading slash.  

As i've said on the issue tracker, we need a better solution here but this will work for now.