Added deltapack to svn.  The deltapack contains all of the platform files for releasing across multiple systems.  This must be a part of the 'target' that is set in your eclipse environment if you want to prepare releases (this has always been the case, now I'm just including it here future devs do not have to go find it).
Added rulebenderPlatform.target.  Eclipse RCP applications require a target platform for execution that defines the plugin environment.  Admittedly I do not know enough about this.  Previously our target has been whatever version of eclipse was running on the current dev machine, and this is not recommended.  The .target file will define a plugin env that is consistent among devs regardless of what eclipse version and plugin environment they are using.  This version of the .target file contains the delta pack and whatever is set in the eclipse installation pointed to by $ECLIPSE_HOME.  So, it effectively is the same as before but I will be updating it once it becomes a higher priority.  At least now we are explicitly setting the target instead of just letting devs figure it out and add their own platform manually.