Contact maps are back, but I had to hack a few things that should be cleaned up later.
Rob attempted to replace creating the bngl ast using the BNGParser jar file with using his bng console process to generate xml, and then produce the same abstract syntax tree as the BNGParser did... but the method to generate the xml is labelled with "//not working yet !!", and the method to generate the ast is commented out entirely.  He left in my old calls to the BNGParser.jar lib, but did not assign the return values to anything and instead used his own calls that instead of returning an ast just returned a File.  This type change was propagated throughout the system, but the handling of it was not changed in any location... Also, his code failed to generate even the xml for any models that I tried. So, commented out his code and changed the filetypes back. 

His bng console class also has an error that prevents the process from being started correctly unless the bng path has been correctly set before the process attempts to start, which is when RB starts.  The workaround is to set the path and then restart RB.  We should really have listeners on the preference values so that the system can respond to changes.  Also, he had a ui blocking call every time an xml file was supposed to be created that lasted 5 seconds wherein the whole app slept while waiting for the xml file to appear. This is a huge impact to usability and will need to be run on it's own thread.

There were also a few differences with BNGParser: (1) the outermost xml tag was changed from 'model' to 'Model', (2) 'ListOfMolecules' was changed to 'ListofMolecules', and (3) the 'type' tag was removed from the 'Rate's.  I just changed the RB code to match what was in the parser.

Somebody should test this to make sure I did not break something else. I have been out of the loop and do not know about the new changes that should be included.