Solving problem where the parser would stop after one error
The problem lied in a missing file inside the BNGParser.jar. The file xml.stg has to be included as part of the distribution
The error message that led to this discovery was getting hidden inside an unrestricted try/except(Exception) statement.

solutions:
Add xml.stg to the BNGParser.jar and upload an updated version of the jar file.
Modify BNGLEditor to have more modular exception handling in its getAST method.