Make a Mosbie design change so that the demo video is more natural.
Until now, Mosbie has had a design that probably was intended to ensure that the user selects a directory for input to Mosbie, rather than a file.  This was accomplished by coding the navigation dialog box so that it showed only directories and no files.  Under this design, my tutorial video would need to start by asking the user to "navigate to that place where you know your files are, but where you unfortunately can't see them."  This seems awkward, so I changed the dialog box to include both directories and files, and in my video, I'll be sure to mention that they are to select a directory and not a file.   (And a later release of Mosbie could have some code that ensures that they select only directories.)