Backing off changes from last commit.
This is a very minor enhancement (not even a defect), and it looks like the change is too dangerous to make at this stage of the development cycle.  In order to view a transfer from an old workspace to a new one, as an upgrade it looks like you just need to uncomment the lines that were commented in this defect, and the lines:

       if (!isCleanWorkspace()) {
          _preferences.put(_KeyBioNetGenRootDir, strBNG);
          Activator.getDefault().getPreferenceStore().setValue("SIM_PATH",PickWorkspaceDialog.getLastSetBioNetGenDirectory());
        }

in OKpressed() should be moved to the top, just after the initialization of  strBNG.  It's an easy change, but much testing is needed.